feat(money): migrate from Double to Decimal end-to-end
Money is now Decimal everywhere — Double on payments + booking fees
+ artist base fees was a P0 in principle. Booking + payment totals,
KPI sums, monthTotal, top-artist tiebreakers all use exact arithmetic
via Decimal. Adds a MoneyFormatter helper with .compact / .full so
'AED 28K' rendering lives in one place instead of being inlined in
six views.

Wire-format conversion (Decimal → Double) only at the PostgREST
insert/update boundary, with a comment explaining why. Wave 5 ceiling
is low millions in any GCC currency so the lossless-Double round-trip
is safe.

DTOs touched: PaymentDTO.amount, ArtistDTO.baseFee, BookingDTO.fee.
Stores touched: PaymentsStore, BookingsStore, AnalyticsStore,
ArtistDetailStore. Views touched: PaymentsView, BookingView,
AvailabilityView, ArtistDashboardView. Tests touched: BookingsStore,
Wave51StoreTests.

+++ RostrPlusPackage/Sources/SupabaseClient/DTO/ArtistDTO.swift
@@ -15,7 +15,7 @@ public struct ArtistDTO: Codable, Hashable, Sendable, Identifiable {
     public let stageName: String?
     public let genre: [String]?
     public let citiesActive: [String]?
-    public let baseFee: Double?
+    public let baseFee: Decimal?
     public let currency: String?
     public let rating: Double?
     public let totalBookings: Int?
